+++ app/Main.hs
@@ -19,16 +19,19 @@ module Main (
 import Control.Monad
 import Control.Monad.Logger
 import Control.Monad.Parallel (MonadParallel (..))
+import Control.Monad.Reader
 
 import           Data.List.Extra    (dropEnd, lower)
-import           Data.Text          (Text)
+import           Data.Maybe         (fromJust, isNothing)
 import qualified Data.Text          as T
 import qualified Data.Text.Encoding as T
 import qualified Data.Text.IO       as T
 import           Data.Time          (defaultTimeLocale, formatTime, getZonedTime)
 
 import Git.Fmt
-import Git.Fmt.Options.Applicative.Parser
+import Git.Fmt.Config  as Config
+import Git.Fmt.Exit
+import Git.Fmt.Options
 
 import Options.Applicative
 
@@ -43,31 +46,31 @@ instance MonadParallel m => MonadParallel (LoggingT m) where
             f' g a b = runLoggingT (f a b) g
 
 main :: IO ()
-main = customExecParser gitFmtPrefs gitFmtInfo >>= \options ->
-    runLoggingT (filter options (handle options)) (if optChatty options == Verbose then verboseLog else log)
+main = do
+    options     <- customExecParser gitFmtPrefs gitFmtInfo
+    let chatty  = optChatty options
 
-filter :: Options -> LoggingT m a -> LoggingT m a
-filter options = filterLogger (\_ level -> level >= minLevel)
-    where
-        minLevel = case optChatty options of
-            Quiet   -> LevelError
-            Default -> LevelInfo
-            Verbose -> LevelDebug
+    flip runLoggingT (log chatty) $ filter chatty $ do
+        mConfigFilePath <- Config.nearestConfigFile "."
+        when (isNothing mConfigFilePath) $ panic (Config.defaultFileName ++ ": not found")
+        config <- Config.readConfig $ fromJust mConfigFilePath
 
-log :: Loc -> LogSource -> LogLevel -> LogStr -> IO ()
-log _ _ level msg = forM_ (T.lines . T.decodeUtf8 $ fromLogStr msg) (T.hPutStrLn h)
-    where
-        h = if level == LevelError then stderr else stdout
+        runReaderT (handle options) config
 
-verboseLog :: Loc -> LogSource -> LogLevel -> LogStr -> IO ()
-verboseLog _ _ level msg = do
+filter :: Chatty -> LoggingT m a -> LoggingT m a
+filter Quiet    = filterLogger (\_ level -> level >= LevelError)
+filter Default  = filterLogger (\_ level -> level >= LevelInfo)
+filter Verbose  = filterLogger (\_ level -> level >= LevelDebug)
+
+log :: Chatty -> Loc -> LogSource -> LogLevel -> LogStr -> IO ()
+log chatty _ _ level msg = do
     timestamp <- formatTime defaultTimeLocale "%F %T.%q" <$> getZonedTime
 
     forM_ (T.lines . T.decodeUtf8 $ fromLogStr msg) $ \line ->
-        T.hPutStrLn h (T.unwords [T.pack $ "[" ++ dropEnd 6 timestamp ++ "]", formatLevel level, line])
+        T.hPutStrLn h $ if chatty == Verbose
+            then T.unwords [T.pack $ "[" ++ dropEnd 6 timestamp ++ "]", formattedLevel, line]
+            else line
     where
-        h = if level == LevelError then stderr else stdout
-
-formatLevel :: LogLevel -> Text
-formatLevel = T.take 6 . (`T.append` "  ") . T.drop 5 . T.pack . lower . show
+        h               = if level == LevelError then stderr else stdout
+        formattedLevel  = T.justifyLeft 6 ' ' . T.drop 5 . T.pack . lower $ show level